chore: render functionality separation
